requirements.txt
GitPython
fire
